refactor: inline avatar component
