+++ src/features/artist/components/ArtistBody/ArtistBody.tsx
@@ -10,12 +10,10 @@ import { ArtistPlaylist } from 'features/artist/components/ArtistPlaylist/Artist
 import { ArtistTopOffers } from 'features/artist/components/ArtistTopOffers/ArtistTopOffers'
 import { ArtistWebMetaHeader } from 'features/artist/components/ArtistWebMetaHeader'
 import { useGoBack } from 'features/navigation/useGoBack'
-import { AccessibilityRole } from 'libs/accessibilityRole/accessibilityRole'
 import { AlgoliaOfferWithArtistAndEan } from 'libs/algolia/types'
 import { capitalizeFirstLetter } from 'libs/parsers/capitalizeFirstLetter'
 import { ensureEndingDot } from 'libs/parsers/ensureEndingDot'
 import { highlightLinks } from 'libs/parsers/highlightLinks'
-import { FastImage } from 'libs/resizing-image-on-demand/FastImage'
 import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
 import { CollapsibleText } from 'ui/components/CollapsibleText/CollapsibleText'
 import { ContentHeader } from 'ui/components/headers/ContentHeader'
@@ -69,7 +67,7 @@ export const ArtistBody: FunctionComponent<Props> = ({
         contentContainerStyle={{ paddingTop: headerHeight }}>
         <ViewGap gap={8}>
           <ViewGap gap={6}>
-            <ArtistHeader name={name} avatarImage={image ? <AvatarImage image={image} /> : null} />
+            <ArtistHeader name={name} avatarImage={image} />
             {capitalizedDescriptionWithDot ? (
               <Description gap={1}>
                 <Typo.BodyAccent>Quelques infos à son sujet</Typo.BodyAccent>
@@ -105,18 +103,3 @@ const ContentContainer = styled(IntersectionObserverScrollView).attrs({
 const Description = styled(ViewGap)(({ theme }) => ({
   marginHorizontal: theme.contentPage.marginHorizontal,
 }))
-
-const StyledImage = styled(FastImage)({
-  width: '100%',
-  height: '100%',
-})
-
-const AvatarImage: FunctionComponent<{
-  image: string
-}> = ({ image }) => (
-  <StyledImage
-    url={image}
-    accessibilityRole={AccessibilityRole.IMAGE}
-    accessibilityLabel="artist avatar" // pourquoi en anglais ?
-  />
-)
